Move construct_pubdata method from utils to L1BatchWithMetadata
